requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appnope==0.1.4
APScheduler==3.11.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
blinker==1.7.0
boto3==1.34.89
botocore==1.34.89
bson==0.5.10
build==1.2.1
cachetools==5.3.3
certifi==2025.1.31
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cycler==0.10.0
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
deep-sort-realtime==1.3.2
deepsort==0.0.5
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
duckduckgo_search==2025.4.4
easydict==1.13
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.110.2
feedparser==6.0.11
filelock==3.13.4
filetype==1.2.0
filterpy==1.4.5
fire==0.6.0
Flask==3.0.3
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
flatbuffers==24.3.25
fonttools==4.51.0
frozendict==2.4.6
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.29.0
google-genai==1.10.0
google_search_results==2.4.2
googleapis-common-protos==1.63.0
grandalf==0.8
grpcio==1.72.0rc1
grpcio-status==1.63.0rc1
h11==0.14.0
hf_transfer==0.1.9
httpcore==1.0.5
httptools==0.6.1
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanfriendly==10.0
hydra-core==1.3.2
idna==2.10
importlib-metadata==7.0.0
importlib_resources==6.4.0
ipdb==0.13.13
ipykernel==6.29.4
ipython==8.23.0
itsdangerous==2.2.0
jax==0.4.26
jaxlib==0.4.26
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kagglehub==0.3.10
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-google-genai==2.1.2
langchain-text-splitters==0.3.8
langgraph==0.3.29
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.30
lap==0.4.0
lxml==5.3.2
Markdown==3.6
markdown-it-py==3.0.0
markdownify==0.12.1
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.11
ml-dtypes==0.4.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
multitasking==0.0.11
mypy-extensions==1.0.0
narwhals==1.33.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.2.4
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.17.3
openai==1.23.6
opencv-contrib-python==4.9.0.80
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.74
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
opt-einsum==3.3.0
orjson==3.10.1
ormsgpack==1.9.1
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parse-utils==1.3.6
parso==0.8.4
peewee==3.17.9
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
posthog==3.5.0
primp==0.14.0
prompt-toolkit==3.0.43
proto-plus==1.26.1
protobuf==4.25.6
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybboxes==0.1.6
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydeck==0.9.1
pygame==2.6.1
Pygments==2.17.2
pymongo==4.6.3
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.36.2
regex==2024.4.16
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
roboflow==1.1.27
rpds-py==0.24.0
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.3
sahi==0.11.15
schedule==1.2.2
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.13.2
sendgrid==6.11.0
sentence-transformers==2.7.0
sentencepiece==0.2.0
sgmllib3k==1.0.0
shapely==2.0.4
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.6
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
starkbank-ecdsa==2.2.0
starlette==0.37.2
streamlit==1.44.1
sympy==1.13.1
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.40.1
typeguard==4.4.2
typer==0.12.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2024.1
tzlocal==5.3.1
ultralytics==8.2.2
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.2
wikipedia==1.4.0
wrapt==1.16.0
xxhash==3.5.0
yarl==1.9.4
yfinance==0.2.55
yolov5==7.0.13
yolov8==0.0.2
zipp==3.18.1
zstandard==0.23.0
